requirements.txt
requests>=2.32
python-dotenv>=1.0
beautifulsoup4>=4.12
pytest>=8.0
markdownify>=1.2.2
